requirements.txt
yt_dlp
pyTelegramBotAPI
python-dotenv